// src/api/generated/models/RequestResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { AppUser } from './AppUser';
import type { Event } from './Event';
export type RequestResponseDTO = {
    id?: number;
    appUser?: AppUser;
    event?: Event;
    status?: boolean;
    date?: string;
};